Move IK and clean
