+++ src/pybullet_helpers/inverse_kinematics.py
@@ -1,0 +1,96 @@
+from typing import Sequence
+
+import numpy as np
+import pybullet as p
+
+from predicators.src.settings import CFG
+from predicators.src.structs import JointsState, Pose3D
+
+
+def pybullet_inverse_kinematics(
+    robot: int,
+    end_effector: int,
+    target_position: Pose3D,
+    target_orientation: Sequence[float],
+    joints: Sequence[int],
+    physics_client_id: int,
+    validate: bool = True,
+) -> JointsState:
+    """Runs IK and returns a joints state for the given (free) joints.
+
+    If validate is True, the PyBullet IK solver is called multiple
+    times, resetting the robot state each time, until the target
+    position is reached. If the target position is not reached after a
+    maximum number of iters, an exception is raised.
+    """
+    # Figure out which joint each dimension of the return of IK corresponds to.
+    free_joints = []
+    num_joints = p.getNumJoints(robot, physicsClientId=physics_client_id)
+    for idx in range(num_joints):
+        joint_info = p.getJointInfo(robot,
+                                    idx,
+                                    physicsClientId=physics_client_id)
+        if joint_info[3] > -1:
+            free_joints.append(idx)
+    assert set(joints).issubset(set(free_joints))
+
+    # Record the initial state of the joints so that we can reset them after.
+    if validate:
+        initial_joints_states = p.getJointStates(
+            robot, free_joints, physicsClientId=physics_client_id)
+        assert len(initial_joints_states) == len(free_joints)
+
+    # Running IK once is often insufficient, so we run it multiple times until
+    # convergence. If it does not converge, an error is raised.
+    convergence_tol = CFG.pybullet_ik_tol
+    for _ in range(CFG.pybullet_max_ik_iters):
+        free_joint_vals = p.calculateInverseKinematics(
+            robot,
+            end_effector,
+            target_position,
+            targetOrientation=target_orientation,
+            physicsClientId=physics_client_id,
+        )
+        assert len(free_joints) == len(free_joint_vals)
+        if not validate:
+            break
+        # Update the robot state and check if the desired position and
+        # orientation are reached.
+        for joint, joint_val in zip(free_joints, free_joint_vals):
+            p.resetJointState(robot,
+                              joint,
+                              targetValue=joint_val,
+                              physicsClientId=physics_client_id)
+        # TODO can this be replaced with get_link_pose?
+        ee_link_state = p.getLinkState(
+            robot,
+            end_effector,
+            computeForwardKinematics=True,
+            physicsClientId=physics_client_id,
+        )
+        position = ee_link_state[4]
+        # Note: we are checking positions only for convergence.
+        if np.allclose(position, target_position, atol=convergence_tol):
+            break
+    else:
+        raise Exception("Inverse kinematics failed to converge.")
+
+    # Reset the joint states to their initial values to avoid modifying the
+    # PyBullet internal state.
+    if validate:
+        for joint, (pos, vel, _, _) in zip(free_joints, initial_joints_states):
+            p.resetJointState(
+                robot,
+                joint,
+                targetValue=pos,
+                targetVelocity=vel,
+                physicsClientId=physics_client_id,
+            )
+    # Order the found free_joint_vals based on the requested joints.
+    joint_vals = []
+    for joint in joints:
+        free_joint_idx = free_joints.index(joint)
+        joint_val = free_joint_vals[free_joint_idx]
+        joint_vals.append(joint_val)
+
+    return joint_vals
